chore: Create global type Zettel, NewZettel, ContentType
