--- client/src/models/Zettel.ts
@@ -1,9 +1,0 @@
-export type Zettel = {
-  id: number;
-  title: string | null;
-  content: string;
-  contentType?: string;
-  tags: string[];
-  createdAt: Date;
-  updatedAt: Date;
-};
