+++ client/src/redux/modules/zettelReducer.ts
@@ -1,5 +1,3 @@
-import { Zettel } from "../../models/Zettel";
-
 const CREATE_ZETTEL = "zettel/CREATE_ZETTEL" as const;
 const DELETE_ZETTEL = "zettel/DELETE_ZETTEL" as const;
 const UPDATE_ZETTEL = "zettel/UPDATE_ZETTEL" as const;
